In order to make the metrics package name clearer, the module and all associated namespaces were changed from 'metrics' to 'codahalemetrics'.
e.g. From "package org.eigengo.monitor.output.metrics" to "package org.eigengo.monitor.output.codahalemetrics"
