+++ output-codahalemetrics/src/test/resources/META-INF/monitor/output.conf
@@ -1,0 +1,6 @@
+org.eigengo.monitor.output.codahalemetrics {
+    registry-class: "org.eigengo.monitor.output.codahalemetrics.DefaultRegistryProvider"
+    naming-class:   "org.eigengo.monitor.output.codahalemetrics.DefaultNameMarshaller"
+    prefix: ""
+    refresh: 5
+}
