--- output-metrics/src/test/resources/META-INF/monitor/output.conf
@@ -1,6 +1,0 @@
-org.eigengo.monitor.output.metrics {
-    registry-class: "org.eigengo.monitor.output.metrics.DefaultRegistryProvider"
-    naming-class:   "org.eigengo.monitor.output.metrics.DefaultNameMarshaller"
-    prefix: ""
-    refresh: 5
-}
